requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
